[fdio] store node name in parent map entry
Identify the node in its parent using its ID rather than its name. This
allows a node's parent and its identifier to have the same lifetime
without breaking the parent's maps by storing the name in the parent;
copying the id is cheap, so it is stored on the node as well as its
entry.

Change-Id: I1a0ecf0b77d14134f7c1387eb98991b3f94ef5c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154941
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>