[usb] Rip unnecessary deps out of the request-fidl.h lib
The lib currently includes banjo deps. This removes the banjo deps.
Further, the logging needs to be supported in both DFv1 and DFv2, so a
usb-fidl-dfv2-compat-logging build target is added.

Lastly, the usb-endpoint lib should have been included in the DDK's
allow-list from the beginning. It as accessing the DDK through
transitive public_deps - which is resolved here.

Bug: b/376724095
Change-Id: I97a92009b70a8ace35c6c765ec9e888bcb017440
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1158990
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>