[ufs] Add support background operations
This CL adds the Background Operations feature and enables Urgent
Background Operations when an Exception Event occurs.
Currently, Background Operations is enabled on active or resume and
disabled on suspend, but it should be changed to enable only on
suspend in the future.

Bug: b/42075643
Test: fx test ufs-tests
Change-Id: Ib2e32f5433c069942a97dcdc3ce0a83b311a3a5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105652
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Jaeyoon Choi <j_yoon.choi@samsung.com>
Reviewed-by: Christopher Anderson <cja@google.com>