[memory] Fix logging when not capturing on level changes
When memory_monitor does not capture digests when memory pressure levels
change, we still need to verify that the next capture time is in line
with the capture rate for the level.

This fixes the issue that regular captures are never started when no
capture on level change is configured.

BUG=375462598

Change-Id: Ic9db7ddc1546f1652598709be3422feaf84314ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164632
Commit-Queue: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Guillaume Chelfi <chelfi@google.com>