[paver] Add FindPartitionTableManager
Note: this is a reland of fxrev.dev/1152401, but I've changed the name
of the method to one which (hopefully) makes its purpose and function
more obvious.

Adds fuchsia.paver.Paver/FindPartitionTableManager, which replaces
fuchsia.paver.Paver/UseBlockDevice.

The old API (UseBlockDevice) put the responsibility of finding the
appropriate block device on clients, but this requires clients to have
access to devfs, which is being removed.

The new approach relies on asking the paver service to find the
appropriate block device, and in turn the paver will ask fshost (see
https://fxrev.dev/1151454).

Client will be ported in follow-up CLs.

Bug: 371060853
Change-Id: I7a3d42564d788ab012c182e50c51e62f0e719050
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162294
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>