[bazel_sdk] rename fuchsia_workspace to rules_fuchsia
This change renames the fuchsia_workspace repository to rules_fuchsia.
This change is the first step in releasing the static rules in their own
bazel repository to avoid excessive downloads.

This change has no functional impact on the SDK as it is currently only
used to bootstrap the repository rules.

Bug: 381104525
Change-Id: I9a5abd689be64bbc6bf4a4b02c6325795f440fca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163917
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>