[zx][rust] Finish boot time support
Add support for the boot reference timeline in:
1. Timer creation
2. Clock creation
3. Port signal packet creation with boot timestamps

The last part required adding the timestamp field to rust's
representation of zx_packet_signal_t (this has been a part of the zircon
packet for a while but was not part of the rust implementation).

This also required updating the fake_clock implementation used in
timekeeper integration tests to allow timers created on the
ZX_CLOCK_BOOT timeline.

Bug: 328306129
Change-Id: I8ee115466d2b65cd677358493e96cd0ce552c892
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164213
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>