[installer] Don't use explicitly specified device
fuchsia.paver.Paver/UseBlockDevices was superceded by
fuchsia.paver.Paver/FindPartitionTableManager.

Note that the installer still expects the user to specify a block
device, but this needs to either be removed at some point (as devfs will
be removed) or the installer itself must be deleted (it is unsupported
at the moment).

Bug: 371060853
Change-Id: Ia04cafa6aa3cb19685913aba3d1300060df5399a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159272
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>