[sestarnix] Revise the permission check & todo-deny audit logs
Introduce an audit callback to the permission-checking helper, that
will be called with an AuditContext structure, that the caller can
choose to add details to before emitting the log.

The todo_check_permission!() helper is updated to use the audit
log hook to emit a track_stub!() report for each denial, while
limiting the number of todo-deny audit logs emitted for a each
location.

Bug: 380399690, 362707360
Change-Id: Ia85baa261c3ef035a7a7d1f66ea3e3220e44bd9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165232
Reviewed-by: Nelly Vouzoukidou <nellyv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>