[bazel] Avoid hashing the same content twice during workspace setup.
The content hash files for `fuchsia_sdk` and `fuchsia_in_tree_idk`
are the same, which resulted in hashing the content of the same IDK
directory twice when this script was running.

As a minor performance improvement, ensure the hashing of the content
only happens once, and create a symlink from one content file to the
other.

NOTE: The `bazel_action.py` script depends on the name of the content
files to match those of the repository that depend on them.

NO_IFTTT=no real need to change toplevel.WORKSPACE.bazel here!

Change-Id: I71ec521900a9e7d26430b80a19783616fc26e084
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165953
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>