[bazel_sdk] Remove a few un-necessary @fuchsia_sdk load() prefixes.
This is a followup of fxr/1163917 which introduced @rules_fuchsia
as the official name of the repository that contains all SDK rules,
without SDK atom targets in it.

This CL removes a few un-necessary use of @fuchsia_sdk// to load
files that are located in the same rules repository.

Change-Id: I13a5dcc546c2491bee2ab549a051348f828b0168
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164756
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>