[virtio] Fix hanging bug during device release.
The fix is to check the IRQ thread exit condition while the thread waits
on further interrupts. Otherwise, the IRQ thread does not terminate
because it waits on further interrupts indefinitely. This bug was
discovered while trying to migrate the virtio-block driver over to DFv2.

Change-Id: I367e8b84618240cdea82e3df2c5dfa2241cdf1b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163993
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>