[aml-uart] Use a single dispatcher
Rather than have separate dispatchers for the wake lease time, fidl, and
interrupts, they were all changed to use the default dispatcher instead.
This allows us to simplify code and remove locks.

Additionally, the manual wake lease acquisition code was changes to use
the fdf_power::WakeLease class. As a result the ability to inject and
modulate a timer was lost. That said, it is still possible to simply
wait out the timer in order to correctly test the functionality as the
timeout value is fairly small it's not a big deal. Some coverage was lost, and
while it could be recovered by injecting inspect and observing the inspect
values, that would really just be testing the WakeLease class and isn't
worthwhile.

Change-Id: Ibb86f46d22ccef61dabc629c466886f8f37a0ec0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164953
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>