[fidl][build] Make SDK target depend on main library
Make the `<fidl_library>_sdk` targets depend on the main
`<fidl_library>` target rather than just the FIDL IR target so that the
additional dependencies in the main target are always added when using
the SDK target.

While the order of the libraries in `all_fidl_json.txt` changes, the set
of libraries remains the same, as verified by comparing the sorted
results from before and after this change.

This does, however, add coverage of additional libraries in a `//vendor`
IDK because it did not have dependencies on the main library target like
those previously in `//sdk/fidl/BUILD.gn`. Some of these libraries fail
linting, so linting of most libraries under `//vendor` must be disabled
until the lint errors can be fixed.

Bug: 381123422, 381163466
Change-Id: I4925fe5417db5ca6b285cd24d40d8cd6051e7b30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163701
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>