[fidl][build] Add dependency on main library to all backends
Make all FIDL backend targets depend on the main FIDL library target
rather than just the IR target so that additional processing is
guaranteed to be executed. Restrict the visibility of the IR target to
prevent future such issues in the future.

This results in the addition of at least 130 libraries to
`all_fidl_json.txt`.

A number of affected libraries have lint warnings, so linting is
temporarily disabled for them.

Bug: 381123422, 295190178, 381096879
Change-Id: Iaf789906fd13e2a4d5964022f02073b260d34433
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163703
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>