[build][elfldltl][ld] Elide unnecessary configs
The //build/config/linux:default-pie config was never appropriate
for non-executable targets.  It wasn't much noticed because for a
long time no shared_library() or loadable_module() targets were
ever used in the Linux-targetted toolchains.  It was actively
worked around in a few places.  This corrects that by adding it
to the set_defaults() list only for executable() targets and not
for the other kinds.

For ld-stub, the no_cpp_standard_library and nolibc configs were
superfluous with using -nostdlib, so no need for them.  The
implicit-host-libs config needs to be removed there as well.

Change-Id: I6120a6cbe37e1bbe72ca75e0713480dfbce9b546
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167492
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>