[debugger] Do not assert state when handling exceptions
When handling exceptions, there's really nothing stopping anything else
in the system from coming along and destroying the process or thread
that caused the exception, this is particularly true when debug_agent is
attaching higher in the job hierarchy instead of directly to a process
since debug_agent is not the first thing seeing the exception. Therefore
we should not make assertions on the state of these objects, but
gracefully handle the error cases instead.

This change is not comprehensive for all of DebugAgent, but should
provide a starting off point.

Change-Id: I5bd604e84c7930cd577d133eff0e9e12c3571c50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155933
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>