[wlan] Ensure PhyManager::create_iface always records an iface creation failure
There was one location where PhyManager::create_iface could fail and not
record the failure as a defect. This change moves the call to
record_defect into create_iface to prevent callers from needing to
remember to call record_defect themselves.

Note that this change will not result in recording a defect in the
previously missed case because PhyManager::add_phy will not insert a phy
and there won't be a place to retain the defect.

Bug: 381908416
Test: Modified existing unit tests to check PhyManager::create_iface
records a defect when a create iface operation fails.
Change-Id: I5ca737c1e77501ff6f23e077027cd7b7d87d9398

Change-Id: Ifc19b0e0cf62ee7501af672e9ae0513b274a9db6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167173
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>