[starnix] Record inspect data per input file
Currently inspect data is only recorded for an open file and is
discarded once the external program holding a reference to that file
decides to close it. This CL implements behavior that maintains a file's
inspect node in the inspect tree when a file is closed by tying the
lifetime of a input file node to it's parent device. We also introduce a
separate inspect node and inspect data for each separate input file.

Bug: b/365571169
Bug: b/370800942
Change-Id: Ifaa8b858f31d3cfeb056ecdcd8f17755f02b97c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164992
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Fuchsia-Auto-Submit: Justin Kobler <jkobler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>