[build][go] Add --fuchsia-source-dir parameter to gen_library_metadata.py
Provide a command-line argument to pass the Fuchsia source directory
to the script directly, instead of trying to auto-detect it.

This is needed to support Ninja sub-builds that are launched from
a build directory that is located under the Bazel workspace directory
(i.e. $NINJA_BUILD_DIR/gen/build/bazel/workspace/) which is itself a
fac-simile of the Fuchsia source directory, but does not include a
`.jiri_manifest` file intentionally (i.e. to avoid loud `jiri` errors).

Without this change, the auto-detection failed in this case with a
strange error message that looks like:

```
ValueError: go_library name must correspond to the source dir: got
   'go.fuchsia.dev/fuchsia/tools/lib/flagmisc', expected
   'go.fuchsia.dev/fuchsia/out/default/gen/build/bazel/workspace/tools/lib/flagmisc'
```

Change-Id: Icc662a055dd4e7079645b3118e919248ec1192c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168337
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: David Turner <digit@google.com>