[sestarnix] Write xattr to root nodes if not already labeled
When mounting a filesystem with xattr labelling we expect either
that it is a pre-existing filesystem, in which case the root node
should already have been labeled, or a new filesystem, in which
case the root node needs to have a label written to it.

Since we don't have clean separation of existing vs new for
cases such as "remotefs" mounts on fxfs storage, where the root
will have been created by Fuchsia before handing it off to Starnix,
we approximate the expected behaviour by attempting to write a
label to the root node if none is found when we init_with_dentry().

Bug: 381117111
Change-Id: I3c55efe443a963f6fbf1f74a00f0043e611f5a4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164812
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>