[netstack3] Introduce SasHandler::get_local_addr_id_for_remote
Add a method to SasHandler to retrieve a strongly-held reference to the
best local address for communicating with a given remote, and implement
the existing get_local_addr_for_remote in terms of it. This will make it
easier for NAT to stash a WeakAddressId, and doesn't affect existing
usages because the AddressId provides easy access to the scalar
IpDeviceAddr.

Bug: 372546692
Change-Id: I0c59951726f797fbbd7ba7f1a1c91afd18c6bc14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164196
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>