[realm_builder] Fix local component generation with dictionaries
RealmBuilder has a feature that auto generates declarations for local
component. When a route in the parent targets the local component as a
child, RealmBuilder will generate the `use` declaration from that, by
copying all of the Capability fields into their component.decl
equivalents.

If the parent route had `from_dictionary` set, this algorithm is
incorrect: if the parent sets `from_dictionary`, that means it has
already extracted the capability from the dictionary, so in the child's
context the capability is just a bare capability, not part of a
dictionary. The correct thing to do in this case is to erase
`from_dictionary` from the child's declaration.

There is a symmetric bug for autogenerated exposes. In theory we could
auto generate a dictionary to put the child's exposed capability in, but
this use case does not seem very important so we've elected to simply
skip auto generation in this instance.

Change-Id: Ie44384d7ec52c72779ed6257ecbdb0aa3c380954
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167032
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Gary Bressler <geb@google.com>