[network] Change the default netstack to Netstack3
This is to catch anything that isn't using the pre-defined products.

Since we don't want to transition terminal yet, we need to
explicitly pin it to Netstack2.

The bad_signature_update test change is required because the image size increases slightly when Netstack3 is used.

Bug: 370546232
Change-Id: I0b1cf2acb37ca27378efd8ffb824fddbf438acdf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128734
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>