[time] Defines a custom time type for the C++ use of the UTC clock
time_utc is not zx::basic_time<1>.  It was defined like so for
expediency when we split off the boot timeline from the monotonic
timeline.

This change ensures that timekeeper::time_utc is a distinct type, but
still equipped with all the algebraic tooling that is available for the
family of types `zx::basic_time<T>`.

Bug: 376107748
Change-Id: I6f0382a5d257e00405415d8740616a72e9d4d206
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167757
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Thomas Gales <tgales@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>