[component][test] AddChildRealmFromDecl
Introduces a new API into the component testing FIDL for adding a child
realm that initializes with a compoennt decl. Very similar to the newly
added AddChildRealmFromRelativeUrl but removes the need for the
declaration to be in a file in the pkg, and be dynamically loadable.

Change-Id: I521c3eb421bfdd007caaaa87639a7d8411cbf603
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168035
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>