[packet-formats] Reinterpret MLDv2/IGMPv3 queries
When operating as an older version listener, if we're capable of
decoding the new version queries we should read them as v1 queries. The
only field that carries a semantic difference there is how to interpret
the response code, which changed in the later versions.

To make the transmutation clearer, introduce methods to reintrepret the
underlying bytes as the older version so netstack3 can operate in strong
older version mode.

Bug: 42071006
Change-Id: I9b2c5c9746ed39069cb7dbbd573c293aa92a892f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168175
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>