[kernel] Fix overflow when updating the platform timer
The timer code did not use the overflow aware zx_ticks_sub_ticks
function when translating from a boot timer deadline to a raw ticks
deadline. This meant that it was possible for the platform timer to
inadvertently be set to an incorrect value, leading to stalls in the
system.

Change-Id: I9026c31b988035fe0ba1103cc5bb7ae5751cf541
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168218
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>