[fxfs] check stream size before updating on allocate
Allocate potentially needs to update the stream size - if the allocation
range is past the current end of the file it's expected to extend it.
When checking that, writes that extended the size of the file could be
sitting in the pager vmo, if it hasn't flushed yet, so instead of
checking the internal handle for the size, we should check the stream
for the most up-to-date size.

Bug: 293943124
Change-Id: I83aef579845570e7d28d8ca2636948dcd2ac26ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169419
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>