[sdk][ctf] Use static driver_test_realm shard
.. in hwinfo and diagnostics CTF test components. The previously
configured child components do not resolve because of the missing dep,
which is obscured if the test's root realm is lazily started. When the
root realm is eagerly started, it becomes a hard error that causes the
test to fail.

Further, adding a dep for the previous shard that was included in the
manifests causes other issues because they expect the tests to interact
with the driver_test_realm, which these tests to not.

Because the requirements for these tests are to simply have a statically
configured topography of children, the dependency is changed to the
static driver_test_realm topography offered through the SDK and the
manifest shard is updated accordingly.

Change-Id: I4170bf6c225371ed1040ec54b4eda490d1724e1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168592
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>