[ffx][errors] Refactoring ffx/lib/errors
This refactors s/d/ffx/lib/errors removing the dependencies on
FIDL. This is done by introducing a new enum, FfxTargetError in its
own crate to handle mapping the FIDL error to a non-FIDL generated error
type so traits can be derived (such as thiserror::Error).

The err member of `FfxError` is changed to be a Boxed trait for
getting the error messasge and the exit code.

This change is being done to improve the build time of the platform,
so that tools that are used in the build do not need to depend
indirectly on libraries such as target, FIDL and related communication
libraries.

The near term benefit will be tools that do not communicate
with the Fuchsia device will not depend on target communication related
libraries, such as the ffx daemon, overnet, et. al.

This is the first refactor in a series to realize this benefit by
pushing down target related error handling into the target crate and
make the high level error library more abstract.

Change-Id: I9aa156044b2c95a1606e674fb59c15617f69c8ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159515
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>