[ffx][repo] Filter daemon repos that aren't running
A previous CL improved repo server tracking for daemon based servers,
especially ones that were in isolate dirs. However, this had the side
effect of listing configured repositories that were not actually running
as part of the daemon process.

This CL filters those repositories based on the status of the daemon
repository server task.

This makes it better, but there is still the possibility of edge cases
if there is a daemon running in another isolate dir.

Bug: 382422026
Change-Id: I9e4b20214dfd4f5edd617a6222a27712f48b5755
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169017
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>