[netstack3] Cache direction in per-packet filter metadata
Currently, both conntrack and NAT logic independently attempt to
determine the direction of a packet with respect to its conntrack
connection. Because this check is stateless (i.e. is performed again at
every NAT hook), if a packet is rewritten due to NAT, it is likely to no
longer match either the original or reply tuple of its connection. This
results in the netstack being unable to correctly perform NAT on it.

Instead of recalculating it each time, have conntrack return the
direction along with the connection itself when a packet's connection is
looked up, and then stash it in the per-packet metadata that persists
throughout the IP layer. This allows NAT to correctly be performed in
all hooks that a packet traverses, and also allows slightly simplifying
some conntrack code.

Fixed: 371017876
Test: netstack3-filter-test
Change-Id: I9224c119aa9062afb4f4fa0304224f703a98a9b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168458
Reviewed-by: Peter DeLong <pdelong@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>