[display][fidl] Store config stamps as in-line values.
Previously, the config stamp in OnDisplayVsync() event was passed
in a FIDL `box`, which is a reference to an out-of-line object.
This added unnecessary complexity to the display coordinator and
engine drivers.

This change removes the `box` from the config stamp field. It's now
stored as an in-line value.

Test: no-op.
Bug: 341156315
Change-Id: I0edfdbbed0e578493b998b416c9d6992234bcebb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170274
API-Review: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>