[memory] Split lazy computation of Inspect output
Split the LazyNode computation of memory_monitor Inspect output in
two parts, one for the fast, already-stored data, and one for the data
that need a full capture first. This ensures that we can output as much
data as we can in Inspect even if we hit a timeout.

BUG=365100782

Change-Id: I33ea16fdb636c26589d2c52722ed88385d09736c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1169832
Commit-Queue: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Guillaume Chelfi <chelfi@google.com>