[debugger] Attach to jobs at ComponentStart
Attaching to jobs during ComponentStarting notifications is critical to
attach as high as possible in the job tree for a particular job_only
filter. If we only attach during process starting events, we will only
ever attach to the immediate parent job of that process. The desired
outcome for a job_only filter is to attach as high in the job tree as
possible to stay out of the way of any descendant processes.

Change-Id: I4a69cad5dd2804c814f2bd07c475eeb378575795
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1157212
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>