[netstack2] Replace connect with r*
We hope to simplify the io2 rights model in the future by removing
explicit connect rights. We also want to make sure that service and
diagnostics directories can be enumerated, and also traversed. To
achieve this, we need to replace the use of "connect" rights in all CML
files with "r*".

These rights are *not* required for connecting to actual service
instances, however using a consistent set of r* rights allows various
client libraries (e.g. fdio) to assume a basic set of directory
operations on a given service directory. Right now this is not possible
as fdio does not use *any* rights, so the current set of rights are not
enforced to begin with.

Bug: 346585458
Change-Id: Ib15f487ca7aff0fe2cc88f1dd3ab080ce0395b1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170096
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>