[starnix] Support futex_wait on CLOCK_REALTIME
Add support for passing CLOCK_REALTIME deadlines to futex_wait in
starnix. Previously, this relied on translating the UTC deadline to a
montonic deadline and waiting, which is fragile and will be buggy once
we start pausing the monotonic clock.

This CL supports waiting on CLOCK_REALTIME deadlines by translating them
to ZX_CLOCK_BOOT deadlines. This is still a bit fragile, but we cannot
directly wait on a UTC deadline so this is the best we can do. Switching
to waiting on ZX_CLOCK_BOOT meant that we could no longer use an
InterruptibleEvent to block, as that struct uses a zx::FutexWait to
perform the wait. Instead, we switch to using a Waiter (which defaults
to using a PortWaiter) to block until either:
1. A boot timer is signalled, or
2. A futex_wake notifies the port.

Bug: 361583830
Change-Id: Ic18ef4a34ed1afc6292f22ed3f01494d9ca26f76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164957
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>