[component_manager] compose a dictionary of dir_entry for service caps
When a service capability is routed, return a dictionary instead of a
dir entry. This dictionary is populated and kept up to date using a
filesystem watcher on the service directory in the capability provider's
outgoing directory. This will allow future commits to perform instance
renaming, filtering, and aggregation on dictionaries.

Change-Id: I9a1a693da82c6f6cdec51645b905cf8c1c644635
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136135
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>