[ffx][emu] Derive emulator mac address from name if possible
When the emulator is started with `--name fuchsia-5254-X-Y`, and X and Y
allow to construct a valid mac address, it will not be generated by
hashing but instead immediately converted from the name. This is helpful
for OTA testing with the `--uefi` flag by ensuring that the target
nodename does not change across reboots.

Fixed: b/383061309
Change-Id: I2f0c2c1ef8703df4de5f147df87b5d535a6a6440
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170873
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>