[sestarnix] Handle Tasks being gone, for Task-labeled FsNodes
Some FsNodes are labeled with an associated Task, whose current
label is the FsNode's effective label. An edge-case exists when an
FsNode is access-checked, that is labeled with a Task that has
already terminated.

We now return the policy's "unlabeled" SID in this case, allowing
more code to rely on all FsNodes having been labeled.

Bug: 381210513
Change-Id: I58feb1ccab858c9086e26df0d05404c25e0659ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1170712
Commit-Queue: Wez <wez@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Przemek Pietrzkiewicz <ppi@google.com>
Reviewed-by: Nathaniel Manista <nathaniel@google.com>