[starnix] Implement MREMAP_DONTUNMAP
For private anonymous memory, MREMAP_DONTUNMAP remaps the mapping while
leaving a special mapping behind. This mapping will generate a pagefault
on each access that will be resolved by either userfaultfd or, if the
range is not registered with one, by zero-filling it. This CL only
handles the latter case.

For memory that is not private anonymous, the behavior is not documented
but seems to be similar to the case where remap is called with source
length of zero.

Test: Mremap.MremapDontUnmap*
Fixed: 297372077

Change-Id: Ic7266fe125f18149600489249f29e3542e159171
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163952
Commit-Queue: Maria Glukhova <mariagl@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>