[fuchsia.sys2] Rename RealmQuery.Open -> DeprecatedOpen
Although RealmQuery.Open was removed at API level 25, there still remain
a lot of usages throughout the codebase, and new instances still keep
cropping up. In an effort to discourage this and to make identifying
the remaining instances easier to find, we rename Open to
DeprecatedOpen.

Likewise, we follow the same pattern for the deprecated StorageAdmin
method OpenComponentStorageById. To improve readability of the FIDL
files, we also move the deprecated/removed protocol functionality to the
end of the protocols they are defined in.

Although this requires rewriting history, this shouldn't be a breaking
change for any clients, since the method was already removed at API
level 25, and all servers are in the platform.

Bug: 331322369
Test: No functional change.
Change-Id: I8f9a05e636a7d4639d2c856222bc56ed94bd0731
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162332
Reviewed-by: Miguel Flores <miguelfrde@google.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Gary Bressler <geb@google.com>